ML/requirements.txt
torch
fastapi
ffmpeg-python
transformers
pydantic[dotenv]